Fixs

+++ StrategyTradeSoft/Tick.cs
@@ -3,11 +3,11 @@ namespace StrategyTradeSoft
     public class Tick
     {
         public DateTime Time { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public string Symbol { get; set; }
         public int Volume { get; set; }
 
-        public Tick(DateTime time, string symbol, int volume, float price)
+        public Tick(DateTime time, string symbol, int volume, double price)
         {
             Time = time;
             Symbol = symbol;
